--- chatgpt_telegram_bot_service/chatgpt_telegram_bot/src/main/java/com/starskrime/chatgpt_telegram_bot/service/ChatGptService.java
@@ -1,4 +1,0 @@
-package com.starskrime.chatgpt_telegram_bot.service;
-
-public interface ChatGptService {
-}
