--- chatgpt_telegram_bot_service/chatgpt_telegram_bot/src/main/java/com/starskrime/chatgpt_telegram_bot/service/impl/ChatGptServiceImpl.java
@@ -1,6 +1,0 @@
-package com.starskrime.chatgpt_telegram_bot.service.impl;
-
-import com.starskrime.chatgpt_telegram_bot.service.ChatGptService;
-
-public class ChatGptServiceImpl implements ChatGptService {
-}
