--- chatgpt_telegram_bot_service/chatgpt_telegram_bot/src/main/resources/application.properties
@@ -1,19 +1,0 @@
-
-# ChatGPT API Configuration
-chatgpt.api.baseurl=https://api.chatgpt.com
-chatgpt.api.keyheader=Authorization
-
-# SQLite Configuration
-spring.jpa.database-platform=com.starskrime.chatgpt_telegram_bot.configuration.SQLDialect
-spring.jpa.hibernate.ddl-auto=update
-
-
-spring.datasource.url = jdbc:sqlite:users.db
-spring.datasource.driver-class-name = org.sqlite.JDBC
-
-spring.datasource.username = admin
-spring.datasource.password = admin
-
-# Telegram Bot Configuration
-telegram.bot.token=6100478302:AAETCh7lMcXvp9VZxsV36f26o88ZohBSBEQ
-telegram.bot.username=ChataGptBot
