+++ chatgpt_telegram_bot_service/src/main/java/com/starskrime/chatgpt_telegram_bot/ChatgptTelegramBotApplication.java
@@ -1,0 +1,13 @@
+package com.starskrime.chatgpt_telegram_bot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ChatgptTelegramBotApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ChatgptTelegramBotApplication.class, args);
+    }
+
+}
