+++ chatgpt_telegram_bot_service/src/main/java/com/starskrime/chatgpt_telegram_bot/configuration/TelegramBotConfiguration.java
@@ -1,0 +1,21 @@
+package com.starskrime.chatgpt_telegram_bot.configuration;
+
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.telegram.telegrambots.bots.DefaultBotOptions;
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.TelegramBotsApi;
+import org.telegram.telegrambots.meta.generics.TelegramBot;
+
+@Configuration
+@Getter
+public class TelegramBotConfiguration {
+    @Value("${telegram.bot.token}")
+    private String botToken;
+
+    @Value("${telegram.bot.username}")
+    private String botUserName;
+
+}
