+++ chatgpt_telegram_bot_service/src/main/java/com/starskrime/chatgpt_telegram_bot/service/TelegramBotService.java
@@ -1,0 +1,4 @@
+package com.starskrime.chatgpt_telegram_bot.service;
+
+public interface TelegramBotService {
+}
